requirements.txt
boto3
chalice
pillow